refactoring preprocessors for multiple fields
